+++ ScriptureGuide/constants.h
@@ -18,6 +18,7 @@
 #define PREFERENCES_FILE "/boot/home/config/settings/scriptureguide/settings"
 #define MODULES_PATH "/boot/home/config/settings/sword/"
 #define WEBSITE_URL "http://www.scripture-guide.org/"
+#define SG_APP_SIGNATURE "application/x-vnd.Scripture-Guide"
 #define SG_MANAGER_SIGNATURE "application/x-vnd.wgp.ScriptureGuideManager"
 
 #define FONTSIZE 12
@@ -48,6 +49,7 @@ const uint32 MENU_FILE_NEW				= 'MFnw';
 const uint32 MENU_FILE_QUIT				= 'MFqu';
 const uint32 MENU_PROGRAM_BOOKMANAGER	= 'MPbm';
 const uint32 MENU_PROGRAM_DICTIONARY	= 'MPdi';
+const uint32 MENU_PROGRAM_VERSELISTS	= 'MPvl';
 const uint32 MENU_PROGRAM_EXPORT_PLAIN		= 'MPe1';
 const uint32 MENU_PROGRAM_EXPORT_TSV		= 'MPe2';
 const uint32 MENU_PROGRAM_EXPORT_MARKDOWN	= 'MPe3';